+++ src/redux/reducers/proposal.js
@@ -152,16 +152,52 @@ const onProposalAnswer = (state: Map, action: Object): Map => {
     .set('isProposalAnswerLoading', false);
 };
 
-const onProposalAnswerLoading = (state: Map): Map => {
+const onProposalAnswerLoading = (state: Map, action: Object): Map => {
+  const {
+    payload: { questionId: referenceId, loading = false }
+  } = action;
+
+  let newState = fromJS({});
+
+  const indexOfListToUpdate = state
+    .get('proposalQuestions')
+    .findIndex(listItem => {
+      return listItem.questionId === referenceId;
+    });
+
+  newState = state.setIn(
+    ['proposalQuestions', indexOfListToUpdate, 'loading'],
+    loading
+  );
+
+  const proposalQuestions = newState.get('proposalQuestions');
+
   return state
-    .set('isProposalAnswerLoading', true)
-    .set('proposalAnswerError', undefined);
+    .set('proposalQuestions', proposalQuestions)
+    .set('isProposalAnswerLoading', loading)
 };
 
 const onProposalAnswerError = (state: Map, action: Object): Map => {
-  const { payload } = action;
+  const { payload: { err, questionId } } = action;
+
+  let newState = fromJS({});
+
+  const indexOfListToUpdate = state
+    .get('proposalQuestions')
+    .findIndex(listItem => {
+      return listItem.questionId === questionId;
+    });
+
+  newState = state.setIn(
+    ['proposalQuestions', indexOfListToUpdate, 'loading'],
+    false
+  );
+
+  const proposalQuestions = newState.get('proposalQuestions');
+
   return state
-    .set('proposalAnswerError', payload)
+    .set('proposalAnswerError', err)
+    .set('proposalQuestions', proposalQuestions)
     .set('isProposalAnswerLoading', false);
 };